server/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
tqdm==4.61.2
transformers
numpy>=1.23.0
rdkit-pypi
torch
firebase-admin==6.1.0
google-cloud-storage==2.10.0
flask_cors
openai
pandas
scikit-learn
plotly
matplotlib
seaborn
py3Dmol
Pillow
